requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
requests==2.32.3
urllib3==2.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
requests==2.32.3
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
anyio==4.8.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
numpy==2.2.2
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.10
pytz==2025.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
